requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
celery[redis]
redis
python-dotenv
openai
numpy